pyproject.toml
[tool.coverage.run]
omit = []

[tool.ruff]
line-length = 100

[tool.ruff.lint.pycodestyle]
max-line-length = 100

[project]
name="bikeaction.org"
version="0.0.0"
requires-python = "==3.13.*"
dependencies = [
    "asgiref>=3.8.1",
    "beautifulsoup4>=4.13.3",
    "celery[redis]>=5.4.0",
    "channels[daphne]>=4.2.0",
    "cmarkgfm>=2024.11.20",
    "discord-py-interactions>=5.14.0",
    "dj-stripe>=2.9.0",
    "django>=5.1.15",
    "django-admin-csvexport>=2.3",
    "django-admin-extra-buttons>=2.1.1",
    "django-allauth[mfa]>=65.6.0",
    "django-anymail[mailgun]>=12.0",
    "django-bleach>=1.0.0",
    "django-cors-headers>=4.7.0",
    "django-dotenv>=1.4.2",
    "django-email-log>=1.4.0",
    "django-environ>=0.12.0",
    "django-htmx>=1.23.0",
    "django-ical>=1.9.2",
    "django-leaflet>=0.31.0",
    "django-markdownfield>=0.11.0",
    "django-markdownify>=0.9.5",
    "django-ordered-model>=3.7.4",
    "django-qr-code>=4.1.0",
    "django-recaptcha>=4.0.0",
    "django-redis>=5.4.0",
    "django-relativity>=0.2.6",
    "django-sesame>=3.2.2",
    "django-sitetree>=1.18.0",
    "django-storages[s3]>=1.14.5",
    "django-tuieditor>=0.2.0",
    "easy-thumbnails>=2.10",
    "geopy>=2.4.1",
    "httpx>=0.28.1",
    "markdown>=3.7",
    "opencv-python-headless>=4.13.0.90",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "pyap2>=0.1.14",
    "pygit2>=1.17.0",
    "pynliner>=0.8.0",
    "redis-lock-py>=1.1.1",
    "reportlab>=4.4.5",
    "sentry-sdk[django]>=2.24.1",
    "shapely>=2.0.7",
    "wagtail>=6.4.1",
    "wagtail-link-block>=1.1.7",
    "wagtail-localize>=1.11.3",
    "wagtail-localize-git>=0.15.0",
    "wagtailset>=1.1.2",
    "whitenoise>=6.9.0",
]

[dependency-groups]
deploy = [
    "gunicorn>=23.0.0",
    "uvicorn>=0.27.0",
]
dev = [
    "django-debug-toolbar",
    "hupper>=1.12.1",
    "jurigged>=0.5.7",
    "ruff>=0.15.0",
]
